Remove unnecessary 'err as i32' casts
In instances where err is already a 32-bit int, 'err as i32' is
an unnecessary cast.

+++ libbpf-rs/src/ringbuf.rs
@@ -93,7 +93,7 @@ impl<'a> RingBufferBuilder<'a> {
 
                 // Handle errors
                 if err != 0 {
-                    return Err(Error::System(err as i32));
+                    return Err(Error::System(err));
                 }
             }
